kk this commit removes circular dependency of midi types and makes the config action blocks import tree shakeable by newer vite version
