+++ src/renderer/runtime/runtime.ts
@@ -20,8 +20,7 @@ import { appSettings } from "./app-helper.store";
 import { add_datapoint } from "../serialport/message-stream.store.js";
 import { logger } from "./runtime.store";
 import { v4 as uuidv4 } from "uuid";
-import { getComponentInformation } from "../lib/_configs";
-import * as CodeBlock from "../config-blocks/CodeBlock.svelte";
+import { getComponentInformation } from "../lib/_config-registry";
 import { appClipboard, ClipboardKey } from "./clipboard.store";
 import { type ActionBlockInformation } from "../config-blocks/ActionBlockInformation";
 import { Runtime } from "./string-table";
@@ -29,7 +28,6 @@ import { Grid } from "../lib/_utils";
 import { type GridConnection } from "../serialport/serialport";
 import { GridRuntimeManager } from "./runtime-manager.store";
 import { user_input } from "./user-input.store";
-import { information as elementNameInformation } from "../config-blocks/ElementName.svelte";
 import { type ProfileCloudTypes } from "../main/panels/profileCloud/ProfileCloud";
 
 class NodeData {
@@ -951,10 +949,7 @@ export class GridEvent extends RuntimeNode<EventData> {
 
     const codeBlock = new GridAction(
       undefined,
-      new ActionData(
-        CodeBlock.information.short,
-        actions.map((action) => action.script).join(" "),
-      ),
+      new ActionData("cb", actions.map((action) => action.script).join(" ")),
     );
 
     if (!codeBlock.isValid()) {
@@ -1375,8 +1370,8 @@ export class GridElement extends RuntimeNode<ElementData> {
 
       const setup = this.findEvent(0);
       const action = setup.actionAt(0);
-      if (action?.short === elementNameInformation.short) {
-        const regex = elementNameInformation.valueRegex;
+      if (action?.short === "sn") {
+        const regex = /self:gen\("([^"]*)"\)/;
         const name = action.script.match(regex)[1];
         this.name = name;
       }
